+++ src/neural_mesh_simplification/losses/edge_crossing_loss.py
