+++ src/neural_mesh_simplification/losses/overlapping_triangles_loss.py
