+++ tests/losses/test_proba_surface_distance.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from neural_mesh_simplification.losses.surface_distance_loss_dgl import ProbabilisticSurfaceDistanceLoss
+from neural_mesh_simplification.losses.surface_distance_loss import ProbabilisticSurfaceDistanceLoss
 
 
 @pytest.fixture
